+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.9.6 (backpublish)
+
+* Fix bug parsing asynchronous suspension gap markers at the end of stack
+  traces. (Also fixed separately in 1.10.0-nullsafety.3)
+* Fix bug parsing asynchronous suspension gap markers at the end of stack
+  traces, when parsing with `Trace.parse` and `Chain.parse`. (Also fixed
+  separately in 1.10.0-nullsafety.6)
+
 ## 1.9.5
 
 * Parse the format for `data:` URIs that the Dart VM has used since `2.2.0`.